feat : refactor code
