+++ server/src/types/product.ts
